backend/requirements.txt
# Backend Requirements

# Web framework
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1

# WebSocket
python-socketio==5.10.0
python-engineio==4.8.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis & Celery
redis==5.0.1
celery==5.3.4

# Authentication
PyJWT==2.8.0
cryptography==41.0.7

# Audio processing
numpy==1.26.2
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1

# ML
scikit-learn==1.3.2
joblib==1.3.2

# AWS SDK
boto3==1.34.10

# Utilities
python-dotenv==1.0.0
psutil==5.9.6
marshmallow==3.20.1

# Production server
gunicorn==21.2.0
gevent==23.9.1
gevent-websocket==0.10.1

# Monitoring
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
